requirements.txt
openai>=1.0.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
jinja2>=3.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
faiss-cpu>=1.7.4
sentence-transformers>=2.0.0
scikit-learn>=1.3.0